requirements.txt
ultralytics
fastapi
opencv-python
uvicorn
easyocr
python-magic
numpy